% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6.R
\name{cmip_consolidate}
\alias{cmip_consolidate}
\title{Concatena los archivos NetCDF de distintos miembros del mismo experimento}
\usage{
cmip_consolidate(files = NULL, base_dir)
}
\arguments{
\item{files}{vector con los archivos a consolidar (por ejemplo, la dalida de \code{\link[=cmip_download]{cmip_download()}}).
Si es \code{NULL}, consolida todos los archivos en \code{base_dir}.}

\item{base_dir}{carpeta raiz de la base de datos. Sólo usado si \code{files} es \code{NULL}.}
}
\value{
Un vector de caracteres con los archivos consolidados.
}
\description{
Concatena los archivos NetCDF de distintos miembros del mismo experimento
}
