% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passwords.R
\name{cmip_default_user}
\alias{cmip_default_user}
\alias{cmip_default_user_get}
\alias{cmip_default_user_set}
\title{Usuario por default}
\usage{
cmip_default_user_set(user = NULL)

cmip_default_user_get()
}
\arguments{
\item{user}{usuario para setear como default (\code{NULL} signfica que no hay usuario por default)}
}
\value{
\code{cmip_default_user_set()} define el usuario por defecto (para la sesión actual)
y devuelve el usuario de forma invisible.

\code{cmip_default_user_get()} devuelve el usuario por default seteado con
\code{cmip_default_user_get()} o el usaurio guardad con \code{cmip_key_set()} en caso de
haber sólo uno.
}
\description{
Usuario por default
}
\section{Functions}{
\itemize{
\item \code{cmip_default_user_set}: Define el usuario por default.

\item \code{cmip_default_user_get}: Obtiene el usuario por default.
}}

