% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip_filter_replicas.R
\name{cmip_filter_replicas}
\alias{cmip_filter_replicas}
\title{Filters replicas based on online servers}
\usage{
cmip_filter_replicas(results)
}
\arguments{
\item{results}{A list of search results from \code{\link[=cmip_search]{cmip_search()}}.}
}
\value{
A data.table.
}
\description{
Checks the online status of data nodes and returns filtered results
so that there is a single replica of each instance that is hosted
on an online node.
}
