% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passwords.R
\name{cmip_key}
\alias{cmip_key}
\alias{cmip_key_remove}
\alias{cmip_key_get}
\alias{cmip_key_set}
\title{Manejo de contraseñas}
\usage{
cmip_key_set(user = NULL, password = NULL)

cmip_key_get(user)
}
\arguments{
\item{user, password}{usuario y contraseña. Si son \code{NULL}, se piden interactivamente (recomendado).}
}
\description{
Funciones para guardar, obtener y eliminar contraseñas de CMIP6
}
\details{
El método recomendado es dejar los valores como \code{NULL}. De esta manera, los datos de
autenticación no quedan guardados en texto plano en la historia de comandos de R
}
\section{Functions}{
\itemize{
\item \code{cmip_key_set}: Guarda una cambinación de usuario/contraseña

\item \code{cmip_key_get}: Lee el password de un usuario.
}}

