% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip-node.R
\name{cmip_node_set}
\alias{cmip_node_set}
\alias{cmip_node_get}
\title{Mange ESGF nodes}
\usage{
cmip_node_set(node = "https://esgf-node.llnl.gov")

cmip_node_get()
}
\arguments{
\item{node}{URL of the node to use.}
}
\description{
Mange ESGF nodes
}
\examples{
old_node <- cmip_node_set("https://esgf-node.ipsl.upmc.fr")

# Restore old node
cmip_node_set(old_node)


}
