% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{cmip_root_set}
\alias{cmip_root_set}
\alias{cmip_root_get}
\title{Set or get the root folder for CMIP data.}
\usage{
cmip_root_set(root, mode = "default")

cmip_root_get()
}
\arguments{
\item{root}{Root folder.}

\item{mode}{Mode that defines fodler permissions. Can be an umask valid string or
the aliases:
\itemize{
\item "default": the default umask returned by \code{Sys.umask(NA)}.
\item "shared": read and write permissions to any user (mode "0000")
\item "private": read and write permissions only for the current user (mode "7777")
}}
}
\description{
Set or get the root folder for CMIP data.
}
