% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip-search.R
\name{cmip_simplify}
\alias{cmip_simplify}
\alias{cmip_unsimplify}
\title{Simplifies the output of searches}
\usage{
cmip_simplify(results)

cmip_unsimplify(data)
}
\arguments{
\item{results}{A list of search results from \code{\link[=cmip_search]{cmip_search()}}.}

\item{data}{A simplifided output from \code{cmip_simplify()}}
}
\description{
Removes a lot of less important columns from the output of
\code{cmip_search()}. The full dataset is stored in the hidden column
\code{full_info}.
Use \code{cmip_unsimplify()} to return to the original format (this is needed fr )
}
