% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip-download.R
\name{cmip_size}
\alias{cmip_size}
\title{Computes the total size of a search result in Mb.}
\usage{
cmip_size(results)
}
\arguments{
\item{results}{A list of search results from \code{\link[=cmip_search]{cmip_search()}}.}
}
\description{
Computes the total size of a search result in Mb.
}
