% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip-urls.R
\name{cmip_urls}
\alias{cmip_urls}
\title{Obtain file server URLs of CMIP data}
\usage{
cmip_urls(results)
}
\arguments{
\item{results}{A list of search results from \code{\link[=cmip_search]{cmip_search()}}.}
}
\value{
character vector of URLs
}
\description{
Finds the URLs that \link{cmip_download} would obtain from a search results object.
}
\details{
If a valid URL is not found the return value has an \code{NA} in its place.

Note that the versions to download have "/fileServer/" in the path, whereas those for the remote
access have "/dodsC/" instead (and can be string-substituted).  This works for some sources we
have seen (GDAL for example can within limitations read remotely from the 'fileServer' URLs, but the
NetCDF library needs the 'dodsC' version, again this a pattern observed on some sources not an authorative
claim).
}
