% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmip6-package.R
\docType{package}
\name{rcmip6-package}
\alias{rcmip6}
\alias{rcmip6-package}
\title{rcmip6: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Elio Campitelli \email{elio.campitelli@cima.fcen.uba.ar} (\href{https://orcid.org/0000-0002-7742-9230}{ORCID})

}
\keyword{internal}
