% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip-available.R
\name{cmip_available}
\alias{cmip_available}
\alias{cmip_available_legacy}
\title{List downloaded models.}
\usage{
cmip_available(root = cmip_root_get())

cmip_available_legacy(root = cmip_root_get())
}
\arguments{
\item{root}{Root folder to download and organise the data.}
}
\value{
A data.frame.
}
\description{
Reads files in your local CMIP6 directory and parses the contents into a
data.frame
}
\details{
\code{cmip_available_legacy()} reads from the older way of storing information
and will only work for files downloaded with version <= 0.0.2 of the package.
}
