% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip-available.R
\name{cmip_available}
\alias{cmip_available}
\title{List downloaded models.}
\usage{
cmip_available(base_dir = cmip_folder_get())
}
\arguments{
\item{base_dir}{Root folder to download and organise the data.}
}
\description{
List downloaded models.
}
