% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6.R
\name{cmip_available}
\alias{cmip_available}
\title{Lista datos de CMIP6 disponibles en una carpeta}
\usage{
cmip_available(base_dir = cmip_folder_get())
}
\arguments{
\item{base_dir}{carpeta raiz de la base de datos}
}
\description{
Lista datos de CMIP6 disponibles en una carpeta
}
