% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip-download.R
\name{cmip_download}
\alias{cmip_download}
\title{Downloads CMIP data}
\usage{
cmip_download(results, base_dir = cmip_folder_get())
}
\arguments{
\item{results}{A list of search results from \code{\link[=cmip_search]{cmip_search()}}.}

\item{base_dir}{Root folder to download and organise tha data.}
}
\value{
A list of files, for now.
}
\description{
Downloads CMIP data
}
