% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6.R
\name{cmip_download}
\alias{cmip_download}
\title{Descarga datos de CMIP6}
\usage{
cmip_download(
  results,
  base_dir = cmip_folder_get(),
  user = cmip_default_user_get(),
  system_config = "",
  force_https = FALSE,
  progress_bar = TRUE
)
}
\arguments{
\item{results}{una lista de resultados devuelta por \code{\link[=cmip_search]{cmip_search()}}}

\item{base_dir}{carpeta raiz de la base de datos de CIP6}

\item{user}{usuario para autenticar (ver \link{cmip_key_set})}

\item{system_config}{comandos de sistema para correr antes de iniciar las descargas
(por ejemplo, para setear el proxy).}

\item{force_https}{usar hptts? (ver Detalles)()}

\item{progress_bar}{Mostrar una barra de progreso?}
}
\value{
Un vector de caracteres con los archivos descargados (que puede pasarse directamente a \code{\link[=cmip_consolidate]{cmip_consolidate()}})
}
\description{
Descarga datos de CMIP6
}
\details{
Los archivos se descargan en una estructura de carpetas estándard:
[base_dir]/Download/Format/Data_used/\{experiment_id\}/\{frequency\}/\{variable_id\}/\{variable_id\}\emph{\{table_id\}}\{source_id\}\emph{\{experiment_id\}\emph{r\{realization_index\}i\{initialization_index\}p\{physics_index\}f\{forcing_index\}}\{grid_label\}}\{datetime_start\}-\{datetime_stop\}.\{ext\}

Algunos modelos fallan al descargarse con el script original porque éste usa http y el servidor
espera https. Con \code{force_https = TRUE} se cambia la URL interna del script para usar https.
}
