% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6.R
\name{cmip_folder_set}
\alias{cmip_folder_set}
\alias{cmip_folder_get}
\title{Define y obtiene la carpeta raiz de los datos CMIP6}
\usage{
cmip_folder_set(base_dir, mode = "default")

cmip_folder_get()
}
\arguments{
\item{base_dir}{carpeta raiz de los datos de CMIP6}

\item{mode}{modo que define los permisos de esa carpeta.
Puede ser un caracter válido de umask o los alias:
\itemize{
\item "default": no hacer ningún cambio
\item "shared": permisos de lectura, escritura y ejecución para todos los usuarios (777)
\item "private": permisos de lectura, escritura y ejecución sólo para el usuario actual (770)
}}
}
\description{
Define y obtiene la carpeta raiz de los datos CMIP6
}
