% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6.R
\name{cmip_search}
\alias{cmip_search}
\alias{cmip_url_to_list}
\title{Busca datos del CMIP6}
\usage{
cmip_search(query)

cmip_url_to_list(url)
}
\arguments{
\item{query}{una lista que define la búsqueda.}

\item{url}{url de búsqueda (ver Detalles)}
}
\value{
Una lista con los resultados de la búsqueda que puede ser pasada a un data.frame con
\code{\link[=as.data.frame]{as.data.frame()}}
}
\description{
Busca datos del CMIP6
}
\details{
La mejor forma de obtener una \code{query} válida es ir al portal de búsqueda
(\url{https://esgf-node.llnl.gov/search/cmip6/}),
realizar una búsqueda que se aproxime
a lo que uno quiere. Debajo del número de resultados hay un link que dice
"return results as JSON", copiar ese link y usar \code{cmip_url_to_list()} para convertir
eso en una lista que luego uno puede modificar. En RStudio, se puede usar
un AddIn para hacerlo más rápido.
}
