% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip6.R
\name{cmip_size}
\alias{cmip_size}
\title{Calcula el tamaño total de una búsqueda en megabites}
\usage{
cmip_size(results)
}
\arguments{
\item{results}{una lista de resultados devuelta por \code{\link[=cmip_search]{cmip_search()}}}
}
\description{
Calcula el tamaño total de una búsqueda en megabites
}
