% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmip6-package.R
\docType{package}
\name{rcmip6-package}
\alias{rcmip6}
\alias{rcmip6-package}
\title{rcmip6: Download CMIP6 Data}
\description{
Search and download CMIP6 data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eliocamp/rcmip6}
  \item Report bugs at \url{https://github.com/eliocamp/rcmip6/issues}
}

}
\author{
\strong{Maintainer}: Elio Campitelli \email{elio.campitelli@cima.fcen.uba.ar} (\href{https://orcid.org/0000-0002-7742-9230}{ORCID})

}
\keyword{internal}
